requirements.txt
python-telegram-bot==13.10
python-dotenv==0.19.2
click==8.0.3
joblib==1.1.0
nltk==3.6.7
regex==2022.1.18
tqdm==4.62.3